closes #24 Moved search to Study API
